backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
crewai>=0.1.0
pandas>=2.2.0
python-dotenv>=1.0.0
langchain>=0.1.0
supabase>=2.3.0
boto3>=1.34.0 